% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_data_reduction_functions.R
\name{blank_cor_hgis}
\alias{blank_cor_hgis}
\title{Blank correct HGIS data}
\usage{
blank_cor_hgis(data, blanks = NULL, fmstd = 1.0398)
}
\arguments{
\item{data}{A data frame of normalized per-sample data.}

\item{blanks}{A vector of positions to use for blanks. Will use target types in \code{data} if not provided.}

\item{fmstd}{The expected fraction modern of the normalizing standard.}
}
\value{
A data frame with blank corrected fraction modern and error.
}
\description{
Blank correct HGIS data
}
