% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{compare_consensus}
\alias{compare_consensus}
\title{Compare data to consensus Fm}
\usage{
compare_consensus(data)
}
\arguments{
\item{data}{A results dataframe with fm_consensus, fm_corr, and sig_fm_corr}
}
\value{
A data frame showing agreement with consensus for all samples.
}
\description{
Compare data to consensus Fm
}
