% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{compare_replicates}
\alias{compare_replicates}
\title{Compare agreement of replicate samples}
\usage{
compare_replicates(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Will use normalized data and error if blank corrected data unavailable.
}
