% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{concCO2}
\alias{concCO2}
\title{Calculate fraction of CO2 in a container vs. time when displaced by another gas}
\usage{
concCO2(t, V = 7000, r = 100, flow = 1)
}
\arguments{
\item{t}{time in same units as r}

\item{V}{volume of vessel in same units as r}

\item{r}{rate of displacement gas flow}

\item{flow}{rate of gas flow in capillary or outflow. Default of 1 returns fraction of CO2. Should be less than r.}
}
\value{
Fractional concentration of CO2 in vessel (and outflow).
}
\description{
CO2 is displaced from a vial by introducing helium through one side
of a double needle and allowing the mixture to flow out via the
other side of the needle. Pressure is maintained at 1ATM as the
outlet is open to air.
}
\details{
TODO: Allow specification of a starting mixture of HE and CO2.
}
