% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_data_reduction_functions.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag outliers using a list of measurements}
\usage{
flag_outliers(data, outliers)
}
\arguments{
\item{data}{A data frame of run data, processed by \code{process_hgis_results}.}

\item{outliers}{A data frame of pos and meas of outliers}
}
\value{
The data frame with outliers flagged as FALSE in ok_calc.
}
\description{
Flag outliers using a list of measurements
}
