% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{hgis_eff}
\alias{hgis_eff}
\title{Title}
\usage{
hgis_eff(ulm, uA, cs = 3)
}
\arguments{
\item{ulm}{Flow in ulm-1.}

\item{uA}{Post-accelerator 12C ion current in uA.}

\item{cs}{Charge state to convert particle-uA to 12C current.}
}
\value{
Efficiency as carbon atoms in vs carbon atoms measured.
}
\description{
Title
}
