% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_database.R
\name{insert_samples}
\alias{insert_samples}
\title{Insert a table of samples into HGIS DB}
\usage{
insert_samples(data, con = hgisdb)
}
\arguments{
\item{data}{A dataframe of sample info in \code{hgis_samples} format.}

\item{con}{A DBI connection object for the HGIS DB}
}
\value{

}
\description{
Insert a table of samples into HGIS DB
}
