% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_data_reduction_functions.R
\name{norm_err}
\alias{norm_err}
\title{Calculate normalization error}
\usage{
norm_err(
  sample,
  standard,
  sample_err,
  standard_err,
  stdrat = 1.0398,
  stdrat_err = 6e-04
)
}
\arguments{
\item{sample}{The normalized Fm of the sample}

\item{standard}{The normalized Fm of the standard}

\item{sample_err}{The error of the sample}

\item{standard_err}{The error of the standard}
}
\value{
The propagated error of the normalized sample
}
\description{
Calculate normalization error
}
