% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_data_reduction_functions.R
\name{norm_hgis}
\alias{norm_hgis}
\title{Normalize HGIS data}
\usage{
norm_hgis(data, standards = NULL)
}
\arguments{
\item{data}{A data frame of per-sample data.}

\item{standards}{A vector of positions to use for standards. Will use target types in \code{data} if not provided.}
}
\value{
A data frame with normalized per-sample data.
}
\description{
Normalize HGIS data
}
