% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hgis_consensus}
\alias{plot_hgis_consensus}
\title{Plot difference from consensus}
\usage{
plot_hgis_consensus(data)
}
\arguments{
\item{data}{A data frame as output from \code{compare_consensus()}}
}
\value{
A ggplot2 object
}
\description{
Plot difference from consensus
}
