% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hgis_summary}
\alias{plot_hgis_summary}
\title{Plot normalized and corrected HGIS summary data}
\usage{
plot_hgis_summary(data)
}
\arguments{
\item{data}{}
}
\value{
A ggplot2 object.
}
\description{
Plot normalized and corrected HGIS summary data
}
