% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{plot_hgis_time}
\alias{plot_hgis_time}
\title{Plot of normalized ratio vs time}
\usage{
plot_hgis_time(data, y_var = normFm, errors = NULL)
}
\arguments{
\item{data}{A tibble in format of output from get_hgis_data().}

\item{y_var}{A column to plot against time}

\item{errors}{A column containing errors for y_var}
}
\value{
A ggplot object with norm ratio and current vs time,
faceted for each pos_name.
}
\description{
Plot of normalized ratio vs time
}
