% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{prestoflow}
\alias{prestoflow}
\title{Pouiselle's equation for flow through small tubing.}
\usage{
prestoflow(dp, r, u, l)
}
\arguments{
\item{dp}{Pressure differential in Pa.}

\item{r}{Inner radius of capillary in m.}

\item{u}{Viscosity in kg/m.}

\item{l}{Length of capillary in m.}
}
\value{
Predicted capillary flow in m^3s-1.
}
\description{
Pouiselle's equation for flow through small tubing.
}
