% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{process_hgis_results}
\alias{process_hgis_results}
\title{Get and process hgis data from results file.}
\usage{
process_hgis_results(file, date, standards)
}
\arguments{
\item{file}{Path to a NOSAMS format AMS results file.}

\item{date}{A vector of dates to subset from file. All data used if missing.}

\item{standards}{A vector of the wheel positions containing standards}
}
\value{

}
\description{
Read data file and perform standard munging.
Normalize using mean of samples marked as "S", or a list of standards.
}
