% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{prop_err}
\alias{prop_err}
\title{Propagate errors}
\usage{
prop_err(err)
}
\arguments{
\item{err}{A vector of errors}
}
\value{
A propagated error for the vector.
}
\description{
Add errors in quadrature for a vector of errors
}
