% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{read_results_file}
\alias{read_results_file}
\title{Read SNICS results file}
\usage{
read_results_file(file)
}
\arguments{
\item{file}{A path to a SNICS results file}
}
\value{
A tibble of results for a wheel
}
\description{
Read SNICS results file
}
