% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{sum_hgis}
\alias{sum_hgis}
\title{Summary table of hgis data}
\usage{
sum_hgis(data)
}
\arguments{
\item{data}{A tibble in format of output from get_hgis_data().}
}
\value{
A tibble with summary data.
}
\description{
Summary table of hgis data
}
