% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgis_data_reduction_functions.R
\name{sum_hgis_targets}
\alias{sum_hgis_targets}
\title{Summarize HGIS data per target}
\usage{
sum_hgis_targets(data, remove_outliers = TRUE)
}
\arguments{
\item{data}{A data frame of run data, processed by \code{get_hgis_data}.}

\item{remove_outliers}{If \code{TRUE}, do not include samples where \code{outlier == TRUE}.}
}
\value{
A data frame of per-sample data.
}
\description{
Summarize HGIS data per target
}
