% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{summarize_standards}
\alias{summarize_standards}
\title{Summarize HGIS standards}
\usage{
summarize_standards(data)
}
\arguments{
\item{data}{An HGIS results dataframe, either raw or summarized by target}
}
\value{
A summary of standard means and errors
}
\description{
Summarize HGIS standards
}
