% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{flowcalc}
\alias{flowcalc}
\title{Calculate flow through a capillary or tube.}
\usage{
flowcalc(pres, ...)
}
\arguments{
\item{pres}{Differential pressure in kPa.}

\item{...}{}
}
\value{
Flow in uL/min.
}
\description{
Calculate flow through a capillary or tube.
}
