% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{get_hgis_data}
\alias{get_hgis_data}
\title{Get and process hgis data from results file.}
\usage{
get_hgis_data(file, date)
}
\arguments{
\item{file}{Path to a NOSAMS format AMS results file.}

\item{date}{A vector of dates to subset from file. All data used if missing.}
}
\value{

}
\description{
Read data file and perform standard munging with amstools::mungeResfile().
Add dilution factor. Normalize using mean of samples marked as "S".
}
