% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGISFunctions.R
\name{plot_hgis}
\alias{plot_hgis}
\title{Boxplot of samples in HGIS test, colored by dilution}
\usage{
plot_hgis(data)
}
\arguments{
\item{data}{A tibble in format of output from get_hgis_data().}
}
\value{
A ggplot object with Boxplots of data.
}
\description{
hline for modern is mean value of tank standard rec 101730
}
