% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{reduce_hgis}
\alias{reduce_hgis}
\title{Reduce HGIS data from an AMS results file}
\usage{
reduce_hgis(
  file,
  date = NULL,
  standards = NULL,
  blanks = NULL,
  outliers = NULL,
  remove_outliers = TRUE,
  get_consensus = TRUE
)
}
\arguments{
\item{file}{A SNICS results file}

\item{date}{Date sample run if analyzing a specific day}

\item{standards}{A vector of standard positions.}

\item{blanks}{A vector of blank positions.}

\item{outliers}{A dataframe of the position and measurement number of outlier runs.}

\item{remove_outliers}{Remove outlier runs from analysis if TRUE}

\item{get_consensus}{Get consensus values for samples if TRUE}
}
\value{
A dataframe of results.
}
\description{
Steps:
\enumerate{
\item Load data
\item Process raw data.
\item Reduce data
\item Normalize
\item Blank Correct
}
}
